image-forensics/image-forensics-master/python_service/requirements.txt
flask==2.0.1
numpy==1.21.0
pillow==8.3.1
requests==2.26.0